fixup for x64

+++ src/Class.h
@@ -96,26 +96,26 @@ namespace LuaBinding {
         }
         static int get_vtable(lua_State *L)
         {
-            lua_pushnumber(L, (double)**(uint32_t**)lua_touserdata(L, 1));
+            lua_pushnumber(L, (double)**(uintptr_t**)lua_touserdata(L, 1));
             return 1;
         }
         static int get_ptr(lua_State* L)
         {
-            lua_pushnumber(L, (double)*(uint32_t*)lua_touserdata(L, 1));
+            lua_pushnumber(L, (double)*(uintptr_t*)lua_touserdata(L, 1));
             return 1;
         }
         static int set_ptr(lua_State *L)
         {
-            *(uint32_t*)lua_touserdata(L, 1) = (uint32_t)lua_tonumber(L, 3);
+            *(uintptr_t*)lua_touserdata(L, 1) = (uintptr_t)lua_tonumber(L, 3);
             return 0;
         }
         static int tostring(lua_State* S)
         {
             char asdf[128];
             if constexpr (sizeof(ptrdiff_t) == 4)
-                snprintf(asdf, sizeof(asdf), "%s{%X}", lua_tostring(S, lua_upvalueindex(1)), (uint32_t)topointer(S, 1));
+                snprintf(asdf, sizeof(asdf), "%s{%X}", lua_tostring(S, lua_upvalueindex(1)), (uintptr_t)topointer(S, 1));
             else if constexpr (sizeof(ptrdiff_t) == 8)
-                snprintf(asdf, sizeof(asdf), "%s{%llX}", lua_tostring(S, lua_upvalueindex(1)), (uint64_t)topointer(S, 1));
+                snprintf(asdf, sizeof(asdf), "%s{%llX}", lua_tostring(S, lua_upvalueindex(1)), (uintptr_t)topointer(S, 1));
             lua_pushstring(S, asdf);
             return 1;
         }
@@ -127,7 +127,7 @@ namespace LuaBinding {
         {
             if (!lua_isuserdata(S, 1) || !lua_isuserdata(S, 2)) lua_pushboolean(S, false);
             else
-                lua_pushboolean(S, *(uint32_t*)lua_touserdata(S, 1) == *(uint32_t*)lua_touserdata(S, 2));
+                lua_pushboolean(S, *(uintptr_t*)lua_touserdata(S, 1) == *(uintptr_t*)lua_touserdata(S, 2));
             return 1;
         }
         static int lua_CIndexFunction(lua_State* S)
@@ -144,9 +144,9 @@ namespace LuaBinding {
                     if (luaL_getmetafield(S, 1, "__name")) {
                         char asdf[128];
                         if constexpr (sizeof(ptrdiff_t) == 4)
-                            snprintf(asdf, sizeof(asdf), "%s{%X}", lua_tostring(S, lua_upvalueindex(1)), (uint32_t)topointer(S, 1));
+                            snprintf(asdf, sizeof(asdf), "%s{%X}", lua_tostring(S, lua_upvalueindex(1)), (uintptr_t)topointer(S, 1));
                         else if constexpr (sizeof(ptrdiff_t) == 8)
-                            snprintf(asdf, sizeof(asdf), "%s{%llX}", lua_tostring(S, lua_upvalueindex(1)), (uint64_t)topointer(S, 1));
+                            snprintf(asdf, sizeof(asdf), "%s{%llX}", lua_tostring(S, lua_upvalueindex(1)), (uintptr_t)topointer(S, 1));
                         lua_pop(S, 1);
 
                         luaL_error(S, "Tried to access invalid %s (property '%s')", asdf, lua_tostring(S, 2));
@@ -779,7 +779,7 @@ namespace LuaBinding {
 
         static int dyn_read(lua_State *L)
         {
-            auto p = (void**)(uint32_t)lua_tonumber(L, 2);
+            auto p = (void**)(uintptr_t)lua_tonumber(L, 2);
             if (p == nullptr || *p == nullptr)
             {
                 lua_pushnil(L);
@@ -796,7 +796,7 @@ namespace LuaBinding {
         }
         static int dyn_cast(lua_State *L)
         {
-            auto p = (void*)(uint32_t)lua_tonumber(L, 2);
+            auto p = (void*)(uintptr_t)lua_tonumber(L, 2);
             if (p == 0)
             {
                 lua_pushnil(L);
